+++ src/interaction/subcommands/moderation/unlock/unlockServer.ts
@@ -4,7 +4,7 @@ import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
 import { GuildAuxdibotCommandData } from '@/interfaces/commands/AuxdibotCommandData';
 import Modules from '@/constants/bot/commands/Modules';
 import { EmbedBuilder, PermissionsBitField } from 'discord.js';
-import { deleteLock } from '@/modules/features/moderation/deleteLock';
+import { deleteLock } from '@/modules/features/moderation/lock/deleteLock';
 import handleLog from '@/util/handleLog';
 import { Log, LogAction } from '@prisma/client';
